Create Persian Gluf
Persian Gluf Token (GLUF) is a vibrant meme coin born from the cultural pride and unstoppable meme energy of the Persian Gulf. Built on the TON blockchain, GLUF combines humor, decentralization, and community-driven spirit to deliver not just a coin—but a movement. With a locked liquidity pool, verified contract, and a roadmap focused on growth, engagement, and creative utility, GLUF aims to ride the global meme coin wave while representing the heart of the Gulf. Whether you're here for the memes or the mission, GLUF welcomes you to surf the tides of innovation.